# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Full Ecommerce REST API",
    "description": "this is full ecommerce rest api with all features ",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "abdelbasset4real@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/v1/users":{
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "get all users (admin - manager)",
        "description": "get all existing users (admin - manager)",
        "operationId": "getUsers",
       
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
        
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user in the system (admin - manager) can be created",
        "description":"Create user in the system (admin - manager) can be created",
        "operationId": "createUser",
        "requestBody": {
          "description":  "Create user in the system (admin - manager) can be created",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
        
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get specific user (admin - manager)",
        "description": "Get specific user by id (admin - manager)",
        "operationId": "specificUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "user not found"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Updates a user (admin - manager) can be updated",
        "description": "",
        "operationId": "updateuser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user that needs to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "update a  user by id in the system",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            }
          },
          "required": true
        },
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Deletes a user (admin-manger)",
        "description": "delete a user by id in the system (admin-manger)",
        "operationId": "deleteProduct",
        "parameters": [
         
          {
            "name": "id",
            "in": "path",
            "description": "user id to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid product value"
          }
        }
        
      }
    },
    "/api/v1/users/changepassword/{id}":{
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "change user password",
        "description":"change user password",
        "operationId": "changePassword",
        "requestBody": {
          "description":  "change user password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/changepassword"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/changepassword"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/changepassword"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/changepassword"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/changepassword"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
        
      }
    },
    "/api/v1/users/getMe":{
      "get": {
        "tags": [
          "Logged user"
        ],
        "summary": "get logged user data",
        "description": "get logged user data",
        "operationId": "getUsers",
       
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
        
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "user": {
        "required": [
          "name",
          "email",
          "password",
          "confirmPassword"
        ],
        "type": "object",
        "properties": {

          "name": {
            "type": "string",
            "example": "admin"
          },
          "email": {
            "type": "string",
            "example": "a@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "06748522558"
          },
          "profileImg": {
            "type": "string",
            "example": "user.png"
          },
          "password": {
            "type": "string",
            "example": "654315645"
          },
          "confirmPassword": {
            "type": "string",
            "example": "654315645"
          },
          "role": {
            "type": "string",
            "example": "user"
          }
        }
      },
      "changepassword":{
        "required": [
          "currentpassword",
          "password",
          "confirmPassword"
        ],
        "type": "object",
        "properties": {

          "currentpassword": {
            "type": "string",
            "example": "123456"
          },
          "password": {
            "type": "string",
            "example": "987654"
          },
          "confirmPassword": {
            "type": "string",
            "example": "987654"
          }
        }
      },
...
    },
...
  }
}